+++ src/httpx2/httpx2/__init__.py
@@ -1,4 +1,5 @@
 from .__version__ import __description__, __title__, __version__
+from ._alias import *
 from ._api import *
 from ._auth import *
 from ._client import *
@@ -16,6 +17,7 @@ __all__ = [
     "__description__",
     "__title__",
     "__version__",
+    "alias_httpx",
     "ASGITransport",
     "AsyncBaseTransport",
     "AsyncByteStream",